+++ packages/gatsby-source-shopify/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.2.19...gatsby-source-shopify@3.2.20) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
 ## [3.2.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.2.18...gatsby-source-shopify@3.2.19) (2020-07-07)
 
 ### Features